Reduce duplication in Google Account data handlers
